+mongoose

+++ src/security/sessions.repository.ts
@@ -1,58 +1,48 @@
-import { ObjectId } from "mongodb";
-import { DeviceAuthSession } from "./session.entity";
-import { sessionsCollection } from "../db/mongoDb";
-import { ISessionDb } from "./session.types";
+import { CustomError } from "../shared/types/error.types";
+import { HttpStatuses } from "../shared/types/httpStatuses";
+import { ObjectId } from "../shared/types/objectId.type";
+import { DeviceSessionModel, SessionDocument } from "./session.entity";
 import { injectable } from "inversify";
 
 @injectable()
 export class SessionsRepository {
-  async saveSession(session: DeviceAuthSession): Promise<ObjectId> {
-    const insertRes = await sessionsCollection.insertOne(session);
-    if (insertRes.acknowledged) {
-      return insertRes.insertedId
-    }
-    throw new Error('Failed to save session')
+
+  async save(session: SessionDocument): Promise<ObjectId> {
+    const res = await session.save();
+    return res._id
   }
 
-  async deleteOtherSessions(lastActiveDate: string, userId: string): Promise<void> {
-    const result = await sessionsCollection
-      .deleteMany({
-        userId,
-        lastActiveDate: { $ne: lastActiveDate }
-      })
+  async deleteOtherSessions(lastActiveDate: Date, userId: string): Promise<void> {
+    const result = await DeviceSessionModel
+      .deleteMany({ userId, lastActiveDate: { $ne: lastActiveDate } })
+      .exec()
     if (result.acknowledged) {
       return;
     }
     throw new Error('Failed to delete sessions, operation not acknowledged by db')
   }
 
-  async deleteSession(lastActiveDate: string): Promise<void> {
-    const result = await sessionsCollection
-      .deleteOne({ lastActiveDate })
+  async deleteSession(session: SessionDocument): Promise<void> {
+    const result = await session.deleteOne()
     if (result.acknowledged) {
       return;
     }
     throw new Error('Failed to delete the session, operation not acknowledged by db')
   }
 
-  async getSession(deviceId: ObjectId, lastActiveDate: string): Promise<ISessionDb | null> {
-    return sessionsCollection
+  async getSession(deviceId: ObjectId, lastActiveDate: Date): Promise<SessionDocument> {
+    const session = await DeviceSessionModel
       .findOne({ _id: deviceId, lastActiveDate })
+      .orFail(
+        new CustomError('Session does not exist or already expired', HttpStatuses.Unauthorized)
+      )
+    return session
   }
+
 
   //WARN: Unsafe without checking iat of the token presenter
-  async getSessionByDeviceId(deviceId: ObjectId): Promise<ISessionDb | null> {
-    return sessionsCollection
+  async findSessionByDeviceId(deviceId: ObjectId): Promise<SessionDocument | null> {
+    return DeviceSessionModel
       .findOne({ _id: deviceId })
-  }
-
-  async refreshSession(deviceId: ObjectId, lastActiveDate: string): Promise<void> {
-    const res = await sessionsCollection.updateOne(
-      { _id: deviceId, },
-      { $set: { lastActiveDate } })
-    if (res.modifiedCount === 1) {
-      return;
-    }
-    throw new Error('Failed to update session, modified count not 1')
   }
 }
